requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
cffi==1.14.6
Django==3.2.6
django-cleanup==5.2.0
django-crispy-forms==1.12.0
django-filter==2.4.0
et-xmlfile==1.1.0
numpy==1.21.2
pandas==1.3.2
psycopg2==2.9.1
pycparser==2.20
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
XlsxWriter==3.0.1
